create confusables util
