+++ app/util/validators/index.js
@@ -1,5 +1,4 @@
 import { ethers } from 'ethers';
-import { confusables } from 'unicode-confusables';
 import Encryptor from '../../core/Encryptor';
 
 export const failedSeedPhraseRequirements = (seed) => {
@@ -34,23 +33,3 @@ export const parseVaultValue = async (password, vault) => {
 export const parseSeedPhrase = (seedPhrase) => (seedPhrase || '').trim().toLowerCase().match(/\w+/gu)?.join(' ') || '';
 
 export const { isValidMnemonic } = ethers.utils;
-
-export const collectConfusables = (ensName) => {
-	const key = 'similarTo';
-	const collection = confusables(ensName).reduce(
-		(total, current) => (key in current ? [...total, current.point] : total),
-		[]
-	);
-	return collection;
-};
-
-const zeroWidthPoints = new Set([
-	'\u200b', // zero width space
-	'\u200c', // zero width non-joiner
-	'\u200d', // zero width joiner
-	'\ufeff', // zero width no-break space
-	'\u2028', // line separator
-	'\u2029', // paragraph separator,
-]);
-
-export const hasZeroWidthPoints = (char) => zeroWidthPoints.has(char);
